Refactor ability system by introducing cooldown and emit offset properties, enhancing release methods for better target handling and adding new release direction and position methods

+++ Modules/Ability/IDamageable.cs
@@ -4,7 +4,7 @@ namespace ModularFramework.Modules.Ability
 {
     public interface IDamageable
     {
-        public void TakeDamage(int damageAmount, DamageType damageType);
+        public void TakeDamage(float damageAmount, DamageType damageType);
 
         public void TakeEffect(IEffect<IDamageable> effect);
